requirements.txt
fastremap==1.13.3
h5py==3.7.0
imageio==2.22.3
kornia==0.6.7
matplotlib==3.6.2
monai==1.0.0
natsort==8.2.0
numba==0.57.1
opencv-python==4.6.0.66
pandas==1.5.1
scikit-image==0.19.3
scipy==1.9.2
seaborn==0.12.1
tifffile==2022.10.10
tqdm==4.64.1
Pillow==10.4.0
numpy==1.24.4
